feat: improve search's fetch chain
When fetching chains from several keywords, mix all chain UIDs before
fetching them by batch in parallel.

The previous behavior was to fetch chains keyword by keyword. This could
explain the performance loss. Indeed, this led to collect all chain UIDs
once in a `HashSet` to mix the chains and then a lookup is needed for
each chain value.

```
search/Searching 1 word time:   [7.6612 ms 7.6811 ms 7.7014 ms]
                        change: [-1.0256% -0.6291% -0.2530%] (p = 0.00 < 0.05)
search/Searching 10 words
                        time:   [7.8668 ms 7.8874 ms 7.9084 ms]
                        change: [+1.3367% +1.7090% +2.0671%] (p = 0.00 < 0.05)
search/Searching 100 words
                        time:   [8.5591 ms 8.5825 ms 8.6062 ms]
                        change: [-1.9081% -1.5165% -1.1034%] (p = 0.00 < 0.05)
search/Searching 1000 words
                        time:   [19.778 ms 19.829 ms 19.881 ms]
                        change: [+13.119% +13.557% +14.003%] (p = 0.00 < 0.05)
```

Note that these benchmarks are run on a desktop and may not be as
precise as a bench run on a server. The trend should be correct though.