fix: use the same fetch function for the search and compact traits
There is no need to make parallel calls to the callback. If there is
some non crypto constraint that requires the number of UIDs fetched not
to be greater than a given number, this should be managed in the
callback code. Findex makes a unique call to the call back at each
search level.