fix: remove calls to `HashMap::remove()`
Rust `HashMap` implementation is based on the SwissTable algorithm,
which is optimized for insert and read operations. Thus for small
values, it may is preferable to `get()` and `clone()` rather than
`remove()`.

The following benchmarks show that removing a `HashMap` entry in order
to get the value stored without copy is about 18% slower than getting
this value and then to copy it.

There seems to be no time difference between small values of sizes 32 or
64B.

```
Deletion
                        time:   [99.916 ns 101.47 ns 103.25 ns]

Get + clone / value length: 32B
                        time:   [84.808 ns 86.263 ns 88.109 ns]

Get + clone / value length: 64B
                        time:   [84.844 ns 86.027 ns 87.410 ns]
```

Plus, I would argue that using `get()` is less prone to introduce errors
when modifying the code since it is not possible to search for a removed
value. It also simplifies the reading.