fix: race condition in `fetch_chains()`
If an upsert modifies the ET and a search is performed before the new
values are inserted in the CT, the search client derives UIDs that do
not exist yet inside the CT and tries fetching them. When trying to
recompose the chain, the removed check would raise an error.

Findex only implements eventual consistency between the two tables, thus
no error should not be raised.