fix: remove failing tests
This is okay since this version is not to be released. It is only the paper's
supporting version. The graph functionality is missing, which changes all
interface arguments (no `IndexedValue`). The tests removed in this commit are to
be re-written upon v7 release.

Also care was taken to write (both sequential and concurrent) tests for each
scheme implemented in this repo. So I have a good confidence about their
correctness.