fix SQLITE_BUSY error upon table creation
For some (unknown) reason, passing the WAL option using the method provided by
the library is slow or sometimes does not release the DB lock, which causes the
subsequent SQL query to fail with an SQLITE_BUSY error.

Even though the proposed patch seems to fix the issue entirely, further
investigation is needed in order to find its cause.