implement a memory batcher for Findex
This memory batcher does not yet implement `g_write` but the necessary structure
is in place to allow for an easy implementation (symmetric with the one of batch
read).

This batcher relies on `tokio` for the channels since the runner needs to be
spawned in a concurrent task to handle the buffer. This contrasts with the
approach in which calls to the different primitive (`b-read`/`g-write`) manage
the buffer. However, these calls need to be launched in concurrent task anyway
and some runtime is there required. Also, caller-managed version is more
sensitive to concurrency bugs: runner-managed version sequentiallizes all events
through the use of a unique runner-challer receiver.

For example in the caller-managed version: if all caller await for the next
answer before all terminating caller process unsubscribe, we reach a deadlock
situation in which all non-terminated caller awaits for an event that will never
come (`buffer.len() == buffer.capacity()`). The runner-managed approach
trivially solves this problem since the runner only stops after the last caller
has unsubscribed.